Refactored Tests Folder
