+++ duesanddos/pyproject.toml
@@ -5,8 +5,40 @@ description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
-    "django>=6.0.2",
-    "pillow>=12.1.1",
+    "asgiref==3.11.1",
+    "black==25.11.0",
+    "boto3==1.34.49",
+    "certifi==2026.2.25",
+    "charset-normalizer==3.4.5",
+    "click==8.1.8",
+    "coverage==7.10.7",
+    "coveralls==4.0.1",
+    "cryptography>=46.0.5",
+    "django==4.2.28",
+    "django-allauth==65.15.0",
+    "django-storages==1.14.4",
+    "docopt==0.6.2",
+    "flake8==7.3.0",
+    "idna==3.11",
+    "jmespath==1.1.0",
+    "mccabe==0.7.0",
+    "mypy-extensions==1.1.0",
+    "oauthlib==3.3.1",
+    "packaging==26.0",
+    "pathspec==1.0.4",
+    "pillow==11.3.0",
+    "platformdirs==4.4.0",
+    "psycopg==3.2.13",
+    "psycopg-binary==3.2.13",
     "psycopg2-binary>=2.9.11",
-    "python-dotenv>=1.2.1",
+    "pycodestyle==2.14.0",
+    "pyflakes==3.4.0",
+    "pyjwt==2.11.0",
+    "python-dotenv==1.2.1",
+    "pytokens==0.4.1",
+    "requests==2.32.5",
+    "sqlparse==0.5.5",
+    "tomli==2.4.0",
+    "typing-extensions==4.15.0",
+    "urllib3==2.6.3",
 ]
